placed lost / found item search from originally main page into two different parts of the app
